LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.9_Automatic_Semicolon_Insertion/S7.9_A6.2_T9-expected.txt

CONSOLE MESSAGE: line 80: SyntaxError: Parse error
S7.9_A6.2_T9

PASS Expected parsing failure

TEST COMPLETE
